Move tests to secondary package
