+++ integration-tests/src/test/kotlin/com/google/devtools/ksp/test/secondary/AndroidDataBindingBuiltInKotlinIT.kt
@@ -1,4 +1,4 @@
-package com.google.devtools.ksp.test.primary
+package com.google.devtools.ksp.test.secondary
 
 import com.google.devtools.ksp.test.fixtures.TemporaryTestProject
 import org.gradle.testkit.runner.GradleRunner
@@ -9,12 +9,11 @@ import org.junit.runner.RunWith
 import org.junit.runners.Parameterized
 
 @RunWith(Parameterized::class)
-class LegacyKaptKspIT(experimentalPsiResolution: Boolean) {
+class AndroidDataBindingBuiltInKotlinIT(experimentalPsiResolution: Boolean) {
     @Rule
     @JvmField
     val project: TemporaryTestProject = TemporaryTestProject(
-        "legacy-kapt",
-        "playground",
+        "android-data-binding-builtinkotlin",
         experimentalPsiResolution = experimentalPsiResolution
     )
 
@@ -27,18 +26,21 @@ class LegacyKaptKspIT(experimentalPsiResolution: Boolean) {
     @Test
     fun testPlaygroundAndroid() {
         val gradleRunner = GradleRunner.create().withProjectDir(project.root)
+
+        // Disabling configuration cache. See https://github.com/google/ksp/issues/299 for details
         gradleRunner.withArguments(
             "clean",
-            ":app:testDebugUnitTest",
+            ":app:assemble",
             "--configuration-cache-problems=warn",
             "--info",
-            "--stacktrace"
-        ).build().let { result ->
-            val output = result.output.lines()
-            val kspTask = output.filter { it.contains(":app:kspDebugKotlin") }
-            val kaptTask = output.filter { it.contains(":app:kaptDebugKotlin") }
-            Assert.assertTrue(kspTask.isNotEmpty())
-            Assert.assertTrue(kaptTask.isNotEmpty())
-        }
+            "--stacktrace",
+        )
+            .build().let { result ->
+                val output = result.output.lines()
+                val kspTask = output.filter {
+                    it.contains(":app:kspDebugKotlin")
+                }
+                Assert.assertTrue(kspTask.isNotEmpty())
+            }
     }
 }
